Fixed enemy spawner, changed some doubles to ints

+++ InvaderLogicLibrary/IHitBox.cs
@@ -8,8 +8,8 @@ namespace InvaderLogicLibrary
 {
     public interface IHitBox
     {
-        double X { get; set; }
-        double Y { get; set; }
+        int X { get; set; }
+        int Y { get; set; }
         double Width { get; set; }
         double Height { get; set; }
